Merge in ~JIANKAIZ/graduation-assignment-jiankai from develop to master

* commit 'c13c69154624037f9f076bb93ad3be661dbcdf30': (64 commits)
  Remove unnecessary casting
  Add new message processor
  Improvements
  Add javadoc
  Remove unused import
  Fix sonar issues
  Fix sonar issues
  Fix sonar issues
  Fix integration tests
  Fix sonarqube issues
  Fix unit test
  Fix unit tests
  Move duplicated classes to a commons module and make modules depend on it
  Added task cancelling and consumer topic offset storing
  Added TaskPackages instead of separate tasks to ensure sequential processing and WIP commit offset storing
  Remove old stuff
  Add javadoc and consumer rebalancer
  Add integration test to worker partition/consumer assignment logic
  Add retry logic on etcd connection and rewatch if one of the watches die
  Add feature adding/removing consumers and distributing them over the available partitions and let workers react to these assignments
  ...

+++ message-queue-kafka/src/main/java/kafka/topic/KafkaTopicConfiguration.java
@@ -4,9 +4,9 @@ import messagequeue.messagebroker.topic.TopicConfiguration;
 
 public class KafkaTopicConfiguration extends TopicConfiguration {
     private final int partitions;
-    private final int replicationFactor;
+    private final short replicationFactor;
 
-    public KafkaTopicConfiguration(String name, int partitions, int replicationFactor) {
+    public KafkaTopicConfiguration(String name, int partitions, short replicationFactor) {
         super(name);
         this.partitions = partitions;
         this.replicationFactor = replicationFactor;
@@ -16,7 +16,7 @@ public class KafkaTopicConfiguration extends TopicConfiguration {
         return partitions;
     }
 
-    public int getReplicationFactor() {
+    public short getReplicationFactor() {
         return replicationFactor;
     }
 }
